requirements.txt
streamlit
Pillow
requests
opencv-python
streamlit-webrtc
ultralytics
